refactor: remove default area sorter and add helper function to sort areas and add a condition for checking areaSorter
