+++ example/src/demos/basic-usage.tsx
@@ -12,23 +12,7 @@ const App = () => {
   const [value, setValue] = useState<CountryPhoneInputValue>({ short: 'US' });
 
   return (
-    <ConfigProvider
-      locale={en}
-      areaSorter={(a, b) => {
-        if (a.name && b.name) {
-          const nameA = a.name.toUpperCase();
-          const nameB = b.name.toUpperCase();
-          if (nameA < nameB) {
-            return -1;
-          }
-          if (nameA > nameB) {
-            return 1;
-          }
-          return 0;
-        }
-        return 0;
-      }}
-    >
+    <ConfigProvider locale={en} areaSorter={true}>
       <CountryPhoneInput
         value={value}
         onChange={(v) => {
